Updated the Pizzeria Palmetto task, whole problems has been resolved
