+++ src/homework_10/Yeghia_Ansuryan/Palmetto/exceptions/PhoneNumberException.java
@@ -1,4 +1,4 @@
-package homework_10.Yeghia_Ansuryan.Palmetto.myExceptions;
+package homework_10.Yeghia_Ansuryan.Palmetto.exceptions;
 
 public class PhoneNumberException extends RuntimeException {
     @Override
