+++ src/homework_10/Yeghia_Ansuryan/Palmetto/product/component/ProductComponent.java
@@ -1,6 +1,7 @@
-package homework_10.Yeghia_Ansuryan.Palmetto.myInterface;
+package homework_10.Yeghia_Ansuryan.Palmetto.product.component;
 
 public interface ProductComponent {
+
     String getName();
 
     double getPrice();
